% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{rres_render_solutions_date}
\alias{rres_render_solutions_date}
\title{Render two versions of a file, with and without solutions}
\usage{
rres_render_solutions_date(input, fext = "html", ...)
}
\arguments{
\item{input}{text: the name of the file to be rendered}

\item{fext}{text for the file extension}

\item{...}{text: other parameters to be passed to the render function}
}
\value{
creates two output files in the default directory. Both have the
current date appended to the file name. One has no solutions in the output
and the second has solutions and the word solutions in the output file name.
}
\description{
This function assumes the YAML header of a .qmd file uses a parameter
called solutions, params: solutions which has two values, TRUE and FALSE.
The file may use this parameter to adjust the R markdown chunk options
such as eval or echo when rendering to output the solutions to exercises.
}
\details{
The function works as a standalone function called in the console or script.
}
\examples{
# rresutils::rres_render_solutions_date("my_path/file")
}
