% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{rres_render_solutions}
\alias{rres_render_solutions}
\title{Render two versions of a file, with and without solutions}
\usage{
rres_render_solutions(input, ...)
}
\arguments{
\item{input}{text: the name of the file to be rendered}

\item{...}{text: other parameters to be passed to the render function}
}
\value{
creates two output files in the default directory. . One has no solutions in the output and the second has solutions and the word solutions in the output file name.
}
\description{
This function assumes the YAML header of an R Markdown file uses a parameter called solutions, params: solutions which has two values, TRUE and FALSE. The file may use this parameter to adjust the r markdown chunk options such as eval or echo when rendering to output the solutions to exercises or not.
}
\details{
The function may be placed in the YAML header of an R Markdown document where it will be invoked on the first output type.

See Rmarkdown Cookbook Section 17.5 as a reference
https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html
}
\examples{
# Use the following YAML setting after removing the leading comment
# knit: rresutils::rres_knit_solutions

}
